Add social security subtraction variable, parameters, unit tests

+++ policyengine_us/variables/gov/states/mn/tax/income/exemptions/mn_exemptions_count.py
@@ -2,7 +2,7 @@ from policyengine_us.model_api import *
 
 
 class mn_exemptions_count(Variable):
-    value_type = float
+    value_type = int
     entity = TaxUnit
     label = "Minnesota exemptions count"
     unit = USD
